lambda-langchain-sample/line_bot/requirements.txt
langchain==0.0.325
line-bot-sdk==3.5.0
urllib3==2.0.7
boto3==1.29.5